# docs/swagger.yaml
definitions:
...
  rabbit.AddParticipantRequest:
    properties:
      created:
        type: integer
      id:
        type: string
      operation:
        type: string
      participant:
        description: кто добавляется в пространство
        type: integer
      space_id:
        type: string
      user_id:
        description: кто добавляет участника
        type: integer
    type: object
...
  rabbit.CreateSpaceRequest:
    properties:
      created:
        type: integer
      id:
        type: string
      name:
        type: string
      operation:
        type: string
      user_id:
        description: создатель пространства
        type: integer
    type: object
...
info:
  contact: {}
  description: 'Веб-сервер, обрабатывающий запросы от Бота Зануды: управление заметками,
    а также перенаправление запросов к другим сервисам (напоминяний, пользователей)'
  title: Веб-сервер для Бота Зануды
paths:
...
  /api/v0/spaces/{space_id}/participants/add:
    post:
      description: Запрос на добавление участника в пространство
      parameters:
      - description: ID пространства
        in: path
        name: space_id
        required: true
        type: string
      - description: |-
          добавить участника в пространство:
          указать айди пользователя,
          айди совместного пространства
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rabbit.AddParticipantRequest'
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Невалидный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Невалидный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Запрос на добавление участника в пространство
  /api/v0/spaces/create:
    post:
      description: Запрос на создание пространства
      parameters:
      - description: |-
          создать пространство:
          указать айди пользователя,
          айди его личного / совместного пространства,
          текст заметки
          тип заметки: текстовый, фото, видео, и т.п.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rabbit.CreateSpaceRequest'
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Невалидный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Невалидный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Запрос на создание пространства
...
swagger: "2.0"